Organized exception files into folders
