+++ src/domain/actor-group/actor-group.service.ts
@@ -8,9 +8,9 @@ import { ActorInput } from '../actor/actor.dto';
 import { ActorGroupInput } from './actor-group.dto';
 import { ActorService } from '../actor/actor.service';
 import { IActor } from '../actor/actor.interface';
-import { EntityNotFoundException } from '../../utils/error-handling/entity.not.found.exception';
+import { EntityNotFoundException } from '../../utils/error-handling/exceptions/entity.not.found.exception';
 import { LogContext } from '../../utils/logging/logging.contexts';
-import { GroupNotInitializedException } from '../../utils/error-handling/group.not.initialized.exception';
+import { GroupNotInitializedException } from '../../utils/error-handling/exceptions/group.not.initialized.exception';
 
 @Injectable()
 export class ActorGroupService {
