+++ src/domain/challenge/challenge.service.ts
@@ -2,11 +2,11 @@ import { Inject, Injectable, LoggerService } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
 import { Entity, Repository } from 'typeorm';
-import { EntityNotFoundException } from '../../utils/error-handling/entity.not.found.exception';
-import { EntityNotInitializedException } from '../../utils/error-handling/entity.not.initialized.exception';
-import { GroupNotInitializedException } from '../../utils/error-handling/group.not.initialized.exception';
-import { RelationshipNotFoundException } from '../../utils/error-handling/relationship.not.found.exception';
-import { ValidationException } from '../../utils/error-handling/validation.exception';
+import { EntityNotFoundException } from '../../utils/error-handling/exceptions/entity.not.found.exception';
+import { EntityNotInitializedException } from '../../utils/error-handling/exceptions/entity.not.initialized.exception';
+import { GroupNotInitializedException } from '../../utils/error-handling/exceptions/group.not.initialized.exception';
+import { RelationshipNotFoundException } from '../../utils/error-handling/exceptions/relationship.not.found.exception';
+import { ValidationException } from '../../utils/error-handling/exceptions/validation.exception';
 import { LogContext } from '../../utils/logging/logging.contexts';
 import { Context } from '../context/context.entity';
 import { ContextService } from '../context/context.service';
