+++ src/domain/ecoverse/ecoverse.resolver.mutations.ts
@@ -3,8 +3,8 @@ import { Resolver } from '@nestjs/graphql';
 import { Args, Context, Mutation } from '@nestjs/graphql/dist/decorators';
 import { GqlAuthGuard } from '../../utils/authentication/graphql.guard';
 import { Roles } from '../../utils/decorators/roles.decorator';
-import { AccountException } from '../../utils/error-handling/account.exception';
-import { BaseException } from '../../utils/error-handling/base.exception';
+import { AccountException } from '../../utils/error-handling/exceptions/account.exception';
+import { BaseException } from '../../utils/error-handling/exceptions/base.exception';
 import { LogContext } from '../../utils/logging/logging.contexts';
 import { Profiling } from '../../utils/logging/logging.profiling.decorator';
 import { ChallengeInput } from '../challenge/challenge.dto';
