+++ src/domain/organisation/organisation.service.ts
@@ -2,7 +2,7 @@ import { Inject, Injectable, LoggerService } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
 import { Repository } from 'typeorm';
-import { EntityNotFoundException } from '../../utils/error-handling/entity.not.found.exception';
+import { EntityNotFoundException } from '../../utils/error-handling/exceptions/entity.not.found.exception';
 import { LogContext } from '../../utils/logging/logging.contexts';
 import { ProfileService } from '../profile/profile.service';
 import { TagsetService } from '../tagset/tagset.service';
