+++ src/domain/reference/reference.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { EntityNotFoundException } from '../../utils/error-handling/entity.not.found.exception';
+import { EntityNotFoundException } from '../../utils/error-handling/exceptions/entity.not.found.exception';
 import { LogContext } from '../../utils/logging/logging.contexts';
 import { ReferenceInput } from './reference.dto';
 import { Reference } from './reference.entity';
