+++ src/domain/relation/relation.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { EntityNotFoundException } from '../../utils/error-handling/entity.not.found.exception';
-import { RelationshipNotFoundException } from '../../utils/error-handling/relationship.not.found.exception';
+import { EntityNotFoundException } from '../../utils/error-handling/exceptions/entity.not.found.exception';
+import { RelationshipNotFoundException } from '../../utils/error-handling/exceptions/relationship.not.found.exception';
 import { LogContext } from '../../utils/logging/logging.contexts';
 import { RelationInput } from './relation.dto';
 import { Relation } from './relation.entity';
