+++ src/utils/error-handling/exceptions/group.not.initialized.exception.ts
@@ -1,8 +1,8 @@
 import { HttpStatus } from '@nestjs/common';
-import { LogContext } from '../logging/logging.contexts';
+import { LogContext } from '../../logging/logging.contexts';
 import { BaseException } from './base.exception';
 
-export class AccountException extends BaseException {
+export class GroupNotInitializedException extends BaseException {
   constructor(error: string, context: LogContext) {
     super(error, context, HttpStatus.BAD_REQUEST.toLocaleString());
   }
